+++ src/pages/RegionPage/RegionPage.jsx
@@ -1,12 +1,13 @@
-import Divider from '../../components/Divider';
-import MainContainer from '../../components/MainContainer';
-import RankTitle from '@components/RankTitle';
-import MyProfile from '../../components/MyProfile';
+import Divider from '@/components/Divider';
+import MainContainer from '@/components/MainContainer';
+import RankTitle from '@/components/RankTitle';
+import MyProfile from '@/components/MyProfile';
 
 import { MyRegionCard, RegionWinnerCard } from '@components/Cards';
 
-import level4 from './../../assets/images/regionLevel/regionLevel4.svg'; //임시
-import RegionRankList from './RegionRankList/components/RegionRankList';
+import level4 from '@/assets/images/regionLevel/regionLevel4.svg'; //임시
+import RegionRankList from '@/pages/RegionPage/components/RegionRankList';
+
 export default function RegionPage() {
 	return (
 		<>
